+++ lib/eddy/summary.rb
@@ -1,0 +1,10 @@
+require "eddy/summary/validate"
+require "eddy/summary/element"
+require "eddy/summary/segment"
+require "eddy/summary/loop"
+require "eddy/summary/transaction_set"
+
+module Eddy
+  # Classes containing data for EDI component definitions.
+  module Summary; end
+end