+++ lib/eddy/summary/transaction_set.rb
@@ -1,7 +1,8 @@
 module Eddy
-  module Schema
+  module Summary
     # An outline of all required components for an EDI transaction set implementation.
-    class TransactionSetSummary
+    class TransactionSet
+
       # A short code identifying the Transaction Set.
       # @return [Integer]
       attr_accessor :id
@@ -12,7 +13,7 @@ module Eddy
       # @return [String]
       attr_accessor :functional_group
       # The components that make up the Transaction Set.
-      # @return [Array<Eddy::Schema::SegmentSummary, Eddy::Schema::LoopSummary>]
+      # @return [Array<Eddy::Summary::Segment, Eddy::Summary::Loop>]
       attr_accessor :components
 
       # @return [void]
@@ -23,7 +24,7 @@ module Eddy
       # @param params [Hash]
       # @return [self]
       def self.create(params = {})
-        summary = TransactionSetSummary.new()
+        summary = new()
         summary.id = params[:id]
         summary.name = params[:name]
         summary.functional_group = params[:functional_group]
@@ -34,9 +35,9 @@ module Eddy
       # @param path [String] Path to a JSON or YAML file containing a valid Segment definition.
       # @return [self]
       def self.from_file(path)
-        raise Eddy::Errors::Error, "Invalid transaction set definition" unless Eddy::Schema.valid_transaction_set_data?(path)
+        raise Eddy::Errors::Error, "Invalid transaction set definition" unless Eddy::Summary.valid_transaction_set_data?(path)
         data = Eddy::Util.read_json_or_yaml(path)
-        return Eddy::Schema::TransactionSetSummary.create(data)
+        return Eddy::Summary::TransactionSet.create(data)
       end
 
       # @param components [Array<Hash>]
@@ -44,9 +45,9 @@ module Eddy
       def process_components(components)
         components.each do |comp|
           if comp.key?(:loop_id)
-            self.components << Eddy::Schema::LoopSummary.create(comp)
+            self.components << Eddy::Summary::Loop.create(comp)
           else
-            self.components << Eddy::Schema::SegmentSummary.create(comp)
+            self.components << Eddy::Summary::Segment.create(comp)
           end
         end
       end
@@ -73,12 +74,12 @@ module Eddy
 
       # Return all components in a single, flattened array.
       #
-      # @return [Array<Eddy::Schema::SegmentSummary, Eddy::Schema::LoopSummary>]
+      # @return [Array<Eddy::Summary::Segment, Eddy::Summary::Loop>]
       def all_components()
         return self.components.map do |comp|
           case comp
-          when Eddy::Schema::LoopSummary    then [comp, comp.all_components()]
-          when Eddy::Schema::SegmentSummary then comp
+          when Eddy::Summary::Loop    then [comp, comp.all_components()]
+          when Eddy::Summary::Segment then comp
           else raise Eddy::Errors::Error
           end
         end.flatten
@@ -86,9 +87,9 @@ module Eddy
 
       # Return one of each kind of loop in the Transaction Set.
       #
-      # @return [Array<Eddy::Schema::SegmentSummary, Eddy::Schema::LoopSummary>]
+      # @return [Array<Eddy::Summary::Segment, Eddy::Summary::Loop>]
       def unique_loops()
-        return self.all_components.select { |c| c.is_a?(Eddy::Schema::LoopSummary) }.uniq(&:id)
+        return self.all_components.select { |c| c.is_a?(Eddy::Summary::Loop) }.uniq(&:id)
       end
 
     end